+++ design-system/button/src/entry-points/loading-button.ts
@@ -1,2 +1,5 @@
 export { default } from '../loading-button';
-export type { LoadingButtonProps } from '../loading-button';
+export type {
+  LoadingButtonProps,
+  LoadingButtonOwnProps,
+} from '../loading-button';