+++ design-system/button/src/entry-points/types.ts
@@ -1,7 +1,10 @@
 // A seperate export which is only types
 export type { Appearance, Spacing, BaseOwnProps } from '../types';
 export type { ButtonProps } from './standard-button';
-export type { LoadingButtonProps } from './loading-button';
+export type {
+  LoadingButtonProps,
+  LoadingButtonOwnProps,
+} from './loading-button';
 export type {
   ThemeTokens,
   ThemeProps,