+++ design-system/logo/src/JiraServiceManagementLogo/index.ts
@@ -1,0 +1,9 @@
+import JiraServiceManagementIcon from './Icon';
+import JiraServiceManagementLogo from './Logo';
+import JiraServiceManagementWordmark from './Wordmark';
+
+export {
+  JiraServiceManagementLogo,
+  JiraServiceManagementIcon,
+  JiraServiceManagementWordmark,
+};